+++ oap-server/analyzer/agent-analyzer/src/main/java/org/apache/skywalking/oap/server/analyzer/provider/trace/UninstrumentedGatewaysConfig.java
@@ -62,7 +62,7 @@ public class UninstrumentedGatewaysConfig extends ConfigChangeWatcher {
         }
         this.settingsString.set(config);
         if (config != null) {
-            onGatewaysUpdated(parseGatewaysFromYml(config));
+            onGatewaysUpdated(parseGatewaysFromYml(""));
         }
     }