+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/ApdexThresholdConfig.java
@@ -90,7 +90,7 @@ public class ApdexThresholdConfig extends ConfigChangeWatcher implements Configu
         }
         rawConfig = config;
         if (config != null) {
-            updateConfig(new StringReader(config));
+            updateConfig(new StringReader(""));
         }
     }